Renamed "EnterCircuitFrame" to "DialogCircuitCode" and moved it to the "fidocadj.dialogs" package
